スピンショットのスクリプトを手直しして、動くようにした。初期化時に噛み合わないため未実装

+++ Assets/Scripts/ECS_Component/GunPorts/SpinGunPortTag.cs
@@ -7,5 +7,5 @@ using Unity.Entities;
 public struct SpinGunPortTag : IComponentData
 {
     // ‰ñ“]‚·‚é‘¬“x
-    public float _spinSpeed;
+    public int _spinSpeed;
 }
